IGNITE-16498 Add ability to add handlers to RestModule from other Ignite modules
