+++ modules/page-memory/pom.xml
@@ -73,6 +73,12 @@
             <groupId>org.apache.ignite</groupId>
             <artifactId>ignite-configuration</artifactId>
             <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.ignite</groupId>
+            <artifactId>ignite-configuration</artifactId>
+            <scope>test</scope>
             <type>test-jar</type>
         </dependency>
     </dependencies>
