+++ pom.xml
@@ -64,6 +64,7 @@
         <module>modules/raft</module>
         <module>modules/raft-client</module>
         <module>modules/rest</module>
+        <module>modules/rest-api</module>
         <module>modules/rocksdb-common</module>
         <module>modules/runner</module>
         <module>modules/schema</module>
